requirements.txt
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
django-filter==25.1

Django==5.2.4
pillow==11.3.0
pytz==2025.2
djongo==1.2.31
pymongo==4.13.2
